+++ CHANGELOG.md
@@ -23,6 +23,8 @@ IMPROVEMENTS:
   [GH-384]
 * `resource/vcd_nsxv_firewall_rule` `rule_tag` must be int to avoid vCD internal exception
   passthrough - [GH-384]
+* `resource/vcd_edge_gateway` and `datasource/vcd_edge_gateway` add `default_network_ip` field to
+  export default edge gateway IP address - [GH-XXX]
 
 BUG FIXES: